+++ CHANGELOG.md
@@ -21,9 +21,9 @@
   Follow the documentation for the examples and migration details.
 * The `group.id`, `config.storage.topic`, `offset.storage.topic`, and `status.storage.topic` fields in `.spec.config` section of the `KafkaConnect` resource are deprecated and will be forbidden in the `v1` CRD API.
   Please use the new `.spec.groupId`, `.spec.configStorageTopic`, `.spec.offsetStorageTopic`, and `.spec.statusStorageTopic` fields instead.
-* The `.spec.connectCluster`, `.spec.clusters`, `.spec.mirrors[].sourceCluster`, `.spec.mirrors[].targetCluster`, and `.spec.mirrors[].heartbeatConnector` are deprecated and will be removed in the `v1` CRD API.
+* The `.spec.connectCluster`, `.spec.clusters`, `.spec.mirrors[].sourceCluster`, `.spec.mirrors[].targetCluster`, and `.spec.mirrors[].heartbeatConnector` fields of the `KafkaMirrorMaker2` resource are deprecated and will be removed in the `v1` CRD API.
   Please use the new fields `.spec.target` and `.spec.mirrors[].source` instead to configure the source and target clusters.
-  If you want to deploy and run the Heartbeat connector, you can use a separate `KafkaConnect` and `KafkaConnector` custom resources.
+  If you want to deploy and run the Heartbeat connector, you can use separate `KafkaConnect` and `KafkaConnector` custom resources.
 
 ## 0.48.0